+++ Ruffles/Channeling/Channels/ReliableSequencedChannel.cs
@@ -126,7 +126,7 @@ namespace Ruffles.Channeling.Channels
             else if (SequencingUtils.Distance(sequence, _incomingLowestAckedSequence, sizeof(ushort)) > 0 && !_receiveSequencer[sequence].Alive)
             {
                 // Alloc payload plus header memory
-                HeapMemory memory = MemoryManager.Alloc(payload.Count - 2);
+                HeapMemory memory = MemoryManager.Alloc((uint)payload.Count - 2);
 
                 // Copy the payload
                 Buffer.BlockCopy(payload.Array, payload.Offset + 2, memory.Buffer, 0, payload.Count - 2);
@@ -153,7 +153,7 @@ namespace Ruffles.Channeling.Channels
             _lastOutboundSequenceNumber++;
 
             // Allocate the memory
-            HeapMemory memory = MemoryManager.Alloc(payload.Count + 4);
+            HeapMemory memory = MemoryManager.Alloc((uint)payload.Count + 4);
 
             // Write headers
             memory.Buffer[0] = HeaderPacker.Pack((byte)MessageType.Data, false);
@@ -271,7 +271,7 @@ namespace Ruffles.Channeling.Channels
                             Sequence = i
                         };
 
-                        connection.SendRaw(new ArraySegment<byte>(_sendSequencer[i].Memory.Buffer, _sendSequencer[i].Memory.VirtualOffset, _sendSequencer[i].Memory.VirtualCount), false);
+                        connection.SendRaw(new ArraySegment<byte>(_sendSequencer[i].Memory.Buffer, (int)_sendSequencer[i].Memory.VirtualOffset, (int)_sendSequencer[i].Memory.VirtualCount), false);
                     }
                 }
             }
